add and config some musure in css
