+++ freelancer-platform/src/components/Navbar.css
@@ -118,7 +118,7 @@
   .navbar-toggle {
     display: block;
     position: absolute;
-    right: 20px;
+    right: 60px;
     font-size: 2em;
     top: 10px;
   }
